feat(swagger): move swagger router to routes/
