+++ routes/swagger/index.js
@@ -9,7 +9,7 @@ const options = {
             version: '1.0.0',
         },
     },
-    apis: ['../routes/*.js'], // files containing annotations as above
+    apis: ['./routes/*.js'], // files containing annotations as above
 };
 
 const openapiSpecification = swaggerJsdoc(options);
